+++ server/src-lib/Hasura/App.hs
@@ -841,9 +841,9 @@ instance (MonadIO m) => HttpLog (PGMetadataStorageAppT m) where
     unLogger logger $ mkHttpLog $
       mkHttpErrorLogContext userInfoM reqId waiReq req qErr Nothing Nothing headers
 
-  logHttpSuccess logger userInfoM reqId waiReq _reqBody _response compressedResponse qTime cType headers () =
+  logHttpSuccess logger userInfoM reqId waiReq reqBody _response compressedResponse qTime cType headers () =
     unLogger logger $ mkHttpLog $
-      mkHttpAccessLogContext userInfoM reqId waiReq compressedResponse qTime cType headers
+      mkHttpAccessLogContext userInfoM reqId waiReq reqBody compressedResponse qTime cType headers
 
 instance (Monad m) => MonadExecuteQuery (PGMetadataStorageAppT m) where
   cacheLookup _ _ _ _ = pure ([], Nothing)